backend/requirements.txt
fastapi
uvicorn
azure-servicebus
azure-identity
pydantic
pydantic-settings
pyodbc